refactor: rename ContextMenu to ActionBar
Based on the current designs, we are no longer using a context menu to
show the actions you can do for a selection.

Instead we are showing the actions as buttons on the bottom of the
screen, so the better name for this component would be an action button
and not a context menu anymore.