feat: just use a simpler dark theme
There are issues with displaying a certain location when a full
css theme is used.

So, this commit changes the code so that it simply sets the theme using
basic overrides instead of selecting a full CSS theme.

The actual epub.js repo must be updated to support bigger and more
complex themes.

For now, the overrides will do.

Note: currently using a hard-coded dark mode. There is way to switch to light mode