feat: remove ability to create account
Since our site is live and we are not yet ready to launch, we don't want
random people to be creating accounts and abusing our system in test.

I created a new branch where all things related to account creation will
be found for testing.

For now, the main branch should have nothing to do with account creation
until we are ready to test with our first users.

Note: this is just done for safety